MAE-636: Reorganize the extension upgrader code
