+++ CRM/MembershipExtras/Upgrader/Steps/Step0007.php
@@ -1,0 +1,10 @@
+<?php
+
+class CRM_MyExtension_Upgrader_Steps_Step0007 {
+
+  public function apply() {
+    $disable = new CRM_MembershipExtras_Setup_Configure_DisableContributionCancelActionsExtension();
+    $disable->apply();
+  }
+
+}
