+++ sql/set_unique_external_ids.sql
@@ -1,5 +1,5 @@
 -- /*******************************************************
--- *Set External IDs as unique fields
+-- * Set External IDs as unique fields
 -- *******************************************************/
 ALTER TABLE civicrm_value_contribution_ext_id ADD CONSTRAINT unique_external_id UNIQUE(external_id);
 ALTER TABLE civicrm_value_contribution_recur_ext_id ADD CONSTRAINT unique_external_id UNIQUE(external_id);
